+++ admin/jobs.php
@@ -27,6 +27,9 @@ $categories = $categoryModel->getAll();
 
 // Handle delete action
 if ($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_POST['action']) && $_POST['action'] === 'delete') {
+    if (!canDeleteContent()) {
+        redirect('jobs.php?forbidden=1');
+    }
     $id = (int)($_POST['id'] ?? 0);
     if ($id) {
         $jobModel->delete($id);
@@ -141,10 +144,12 @@ include 'includes/header.php';
                                        class="text-green-600 hover:text-green-800" title="Edit">
                                         <i class="fas fa-edit"></i>
                                     </a>
+                                    <?php if (canDeleteContent()): ?>
                                     <button onclick="deleteJob(<?= $job['id'] ?>, '<?= addslashes($job['title']) ?>')"
                                             class="text-red-600 hover:text-red-800" title="Delete">
                                         <i class="fas fa-trash"></i>
                                     </button>
+                                    <?php endif; ?>
                                 </div>
                             </td>
                         </tr>
